pyproject.toml
[project]
name = "contract-mas"
version = "0.1.0"
description = "Multi-Agent Contract Analysis System for Contract Lifecycle Management"
authors = [{ name = "Trung Nguyen", email = "trung.nguyen@tum.de" }]
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

dependencies = [
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-anthropic>=0.3.0",
    "langfuse>=2.0.0",
    "anthropic>=0.40.0",
    "datasets>=3.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "scipy>=1.14.0",
    "numpy>=2.0.0",
    "pandas>=2.2.0",
    "tqdm>=4.66.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "types-PyYAML>=6.0.0",
    "types-tqdm>=4.66.0",
    "ipykernel>=6.29.0",
    "jupyter>=1.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --tb=short"
